chore: create rubricFactory module
- consolidates logic for generating rubric-related objects
- added frontend functions to global type defintion
- moved helper functions to modules within utils directory
- TG-216
