+++ palette-types/tsconfig.json
@@ -6,7 +6,7 @@
     "sourceRoot": "./",
     "rootDir": "./",
     "strict": true,
-    "sourceMap": true
+    "sourceMap": true,
   },
   "include": ["src/**/*.ts", "src/**/*.d.ts"]
 }
